+++ electron-react/src/components/FileExplorer/DirectoryComponent.jsx
