+++ electron-react/src/components/context/ShowUIContext.jsx
@@ -1,4 +1,4 @@
-import React, { useState, createContext } from "react";
+import React, { useState, createContext } from 'react';
 
 export const ShowUIContext = createContext();
