+++ electron-react/src/components/framerMotion/AthenaLogo.jsx
