+++ electron-react/src/components/framerMotion/AthenaSvg.jsx
